+++ config/test.exs
@@ -3,7 +3,7 @@ use Mix.Config
 # We don't run a server during test. If one is required,
 # you can enable the server option below.
 config :td_dd, TdDdWeb.Endpoint,
-  http: [port: 3005],
+  http: [port: 4005],
   server: true
 
 # Hashing algorithm just for testing porpouses
@@ -30,18 +30,17 @@ config :td_dd, :auth_service,
   auth_port: "4001",
   auth_domain: ""
 
-config :td_dd, :elasticsearch,
-  search_service: TdDd.Search.MockSearch,
-  es_host: "elastic",
-  es_port: 9200
-
 config :td_dd, permission_resolver: TdDd.Permissions.MockPermissionResolver
 config :td_dd, index_worker: TdDd.Search.MockIndexWorker
-
-config :td_cache, redis_host: "redis"
 
 config :td_dd, :audit_service,
   api_service: TdDdWeb.ApiServices.MockTdAuditService,
   audit_host: "localhost",
   audit_port: "4007",
   audit_domain: ""
+
+config :td_dd, TdDd.Search.Cluster, api: TdDd.ElasticsearchMock
+
+config :td_cache, redis_host: "redis"
+
+config :td_cache, :event_stream, streams: []