+++ lib/td_dd/search/cluster.ex
@@ -1,0 +1,5 @@
+defmodule TdDd.Search.Cluster do
+  @moduledoc "Elasticsearch cluster configuration for TdDd"
+
+  use Elasticsearch.Cluster, otp_app: :td_dd
+end