+++ test/td_dd_web/controllers/data_structure_controller_test.exs
@@ -7,6 +7,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
   alias TdCache.TemplateCache
   alias TdDd.DataStructures
   alias TdDd.DataStructures.DataStructure
+  alias TdDd.DataStructures.PathCache
   alias TdDd.Permissions.MockPermissionResolver
   alias TdDdWeb.ApiServices.MockTdAuditService
   alias TdDdWeb.ApiServices.MockTdAuthService
@@ -59,6 +60,7 @@ defmodule TdDdWeb.DataStructureControllerTest do
     start_supervised(MockTdAuthService)
     start_supervised(MockTdAuditService)
     start_supervised(MockPermissionResolver)
+    start_supervised(PathCache)
     :ok
   end
 
@@ -190,17 +192,12 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
     test "search_all", %{conn: conn, data_structure: %DataStructure{id: id}} do
       conn = post(conn, Routes.data_structure_path(conn, :search), %{})
-      json = json_response(conn, 200)
-      data = json["data"]
-      filters = json["filters"]
 
-      assert length(data) == 1
-
-      template_field_values = Map.get(filters, "field")
-      item = Enum.at(data, 0)
+      assert json = json_response(conn, 200)
+      assert [item] = json["data"]
+      assert filters = json["filters"]
 
       assert Map.get(item, "id") == id
-      assert template_field_values == ["1"]
     end
   end