+++ test/td_dd_web/controllers/metadata_controller_test.exs
@@ -2,6 +2,7 @@ defmodule TdDdWeb.MetadataControllerTest do
   use TdDdWeb.ConnCase
   use PhoenixSwagger.SchemaTest, "priv/static/swagger.json"
 
+  alias TdDd.DataStructures.PathCache
   alias TdDd.Loader.LoaderWorker
   alias TdDd.Permissions.MockPermissionResolver
   alias TdDd.Search.MockIndexWorker
@@ -13,6 +14,8 @@ defmodule TdDdWeb.MetadataControllerTest do
     start_supervised(MockTdAuditService)
     start_supervised(MockTdAuthService)
     start_supervised(MockPermissionResolver)
+    start_supervised(LoaderWorker)
+    start_supervised(PathCache)
     :ok
   end
 
@@ -50,7 +53,7 @@ defmodule TdDdWeb.MetadataControllerTest do
       assert response(conn, 202) =~ ""
 
       # waits for loader to complete
-      LoaderWorker.ping()
+      LoaderWorker.ping(20_000)
 
       search_params = %{ou: "Truedat"}
       conn = get(conn, Routes.data_structure_path(conn, :index, search_params))