+++ test/td_dd_web/controllers/system_controller_test.exs
@@ -3,6 +3,7 @@ defmodule TdDdWeb.SystemControllerTest do
   use PhoenixSwagger.SchemaTest, "priv/static/swagger.json"
 
   alias TdCache.TaxonomyCache
+  alias TdDd.DataStructures.PathCache
   alias TdDd.Permissions.MockPermissionResolver
   alias TdDd.Systems
   alias TdDd.Systems.System
@@ -23,6 +24,7 @@ defmodule TdDdWeb.SystemControllerTest do
     start_supervised(MockTdAuditService)
     start_supervised(MockTdAuthService)
     start_supervised(MockPermissionResolver)
+    start_supervised(PathCache)
     :ok
   end
 
@@ -175,11 +177,8 @@ defmodule TdDdWeb.SystemControllerTest do
     @tag authenticated_user: @admin_user_name
     test "will not break when structure has no versions", %{conn: conn, system: system} do
       insert(:data_structure, system_id: system.id, external_id: "parent")
-
       conn = get(conn, Routes.system_data_structure_path(conn, :get_system_structures, system))
-      data = json_response(conn, 200)["data"]
-
-      assert Enum.empty?(data)
+      assert json_response(conn, 200)["data"] == []
     end
 
     @tag authenticated_no_admin_user: "user"