+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-1721] Reindex automatically when a template changes
+  - Breaking change: New environment variable ES_URL replaces existing ES_HOST/ES_PORT
+
 ## [3.7.0] 2019-09-30
 
 ### Added